move model to mars_lib
